Tests NameSpaces and clases renamed
